+++ src/renderer/src/components/chat.tsx
@@ -6,7 +6,7 @@ import {Messages} from "@/components/messages";
 import {useChat} from "@ai-sdk/react";
 import {IpcChatTransport} from "@/chat-transport";
 import {MultimodalInput} from "@/components/multimodal-input";
-import {useState} from "react";
+import {useEffect, useState} from "react";
 
 export function Chat({
                          id,
@@ -21,7 +21,7 @@ export function Chat({
     const [input, setInput] = useState<string>('');
     const [attachments, setAttachments] = useState<Array<Attachment>>([]);
     const [forceScrollToBottom, setForceScrollToBottom] = useState(false);
-
+    const [stillAnswering, setStillAnswering] = useState(false);
     const {
         messages,
         setMessages,
@@ -45,6 +45,17 @@ export function Chat({
         },
     });
 
+    // Track if the model is still answering based on the status
+    useEffect(() => {
+        console.log("status in chat useEffect:", status);
+        if (status === 'ready' || status === 'error') {
+            setStillAnswering(false);
+        } else if (status === 'submitted' || status === 'streaming') {
+            setStillAnswering(true);
+        }
+    }, [status]);
+
+
     const showSuggestedActions = messages.length === 0 && attachments.length === 0;
 
     return (
@@ -67,6 +78,9 @@ export function Chat({
                     regenerate={regenerate}
                     isReadonly={false}
                     isArtifactVisible={false}
+                    stillAnswering={stillAnswering}
+                    forceScrollToBottom={forceScrollToBottom}
+                    setForceScrollToBottom={setForceScrollToBottom}
                 />
 
                 <form className="flex mx-auto px-4 bg-background pb-4 md:pb-6 gap-2 w-full md:max-w-3xl">
@@ -78,11 +92,10 @@ export function Chat({
                         stop={stop}
                         attachments={attachments}
                         setAttachments={setAttachments}
-                        messages={messages}
                         setMessages={setMessages}
                         sendMessage={sendMessage}
                         showSuggestedActions={showSuggestedActions}
-                        forceScrollToBottom={forceScrollToBottom}
+                        stillAnswering={stillAnswering}
                         setForceScrollToBottom={setForceScrollToBottom}
                     />
                 </form>
